Handle return value of nice(3) properly
The man page for nice(3) states:

    Upon successful completion, nice() shall return the new nice
    value -{NZERO}. Otherwise, -1 shall be returned, the process'
    nice value shall not be changed, and errno shall be set to
    indicate the error.

i.e., negative values that are not -1 are perfectly valid return codes.
(Admittedly, this is a problem with the design of nice(3), but it's what
we've got to work with.)

The only case where nice(3) should fail is with permissions problems.
Since this only affects scheduling priorities (and the set*id calls
should fail with EPERM before we get to the nice call) skipping error
checking should be acceptable.

This fixes a bug I ran into when tmate-slave process was running at
priority -10, and the `nice(1)` call incremented that to -9.